% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual_returns.R
\name{annual_returns}
\alias{annual_returns}
\title{Calculate calendar year returns of a security}
\usage{
annual_returns(x, start.date, end.date = Sys.Date())
}
\arguments{
\item{x}{A Bloomberg security id.}

\item{start.date}{A date.}

\item{end.date}{A date.}
}
\value{
A tibble.
}
\description{
Calculate calendar year returns of a security
}
