% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_model_agg_crossover.R
\name{check_model_agg_crossover}
\alias{check_model_agg_crossover}
\title{Check for model_agg crossovers between different RB series'}
\usage{
check_model_agg_crossover(data)
}
\arguments{
\item{data}{cleaned orion platform tibble}
}
\value{
list with three components: results_all, results_pass, results_fail
}
\description{
Check for model_agg crossovers between different RB series'
}
\examples{
# Load and prep data
aim <- readr::read_csv("Orion Platform - XXXX.XX.csv") |>
    clean_orion_platform()

# Check for violations
crossover_results <- check_model_agg_crossover(aim)
}
