% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_classification_wb.R
\name{create_classification_wb}
\alias{create_classification_wb}
\title{Create a classification workbook to audit Orion products.}
\usage{
create_classification_wb(all_local, include_predictions = TRUE)
}
\arguments{
\item{all_local}{a CSV exported from the Orion Local Products page or transformed Orion Query 10635.}

\item{include_predictions}{logical, whether to include prediction columns. Defaults to TRUE.}
}
\value{
an XLSX workbook
}
\description{
Create a classification workbook to audit Orion products.
}
