% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dated_filename.R
\name{dated_filename}
\alias{dated_filename}
\title{Generate filename with date}
\usage{
dated_filename(x, type)
}
\arguments{
\item{x}{root name of the file}

\item{type}{file extension}
}
\value{
a string
}
\description{
Generate filename with date
}
