% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_equity_beta.R
\name{expected_equity_beta}
\alias{expected_equity_beta}
\title{Calculate the expected equity beta of a portfolio}
\usage{
expected_equity_beta(allocation, market = "AC World Equity")
}
\arguments{
\item{allocation}{The allocation of a portfolio}

\item{market}{The relative market used to calculate equity beta}
}
\value{
a numeric
}
\description{
Calculate the expected equity beta of a portfolio
}
\examples{
x <- get_portfolio_allocation("Market Series ETF 080")
expected_equity_beta(x)

}
