% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_return.R
\name{expected_return}
\alias{expected_return}
\title{Calculate the expected return for a portfolio}
\usage{
expected_return(allocation)
}
\arguments{
\item{allocation}{The asset allocation of a portfolio}
}
\value{
numeric
}
\description{
Calculate the expected return for a portfolio
}
\examples{
x <- "Market Series ETF 070"
y <- get_portfolio_allocation(x)
expected_return <- expected_return(y)

}
