% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_risk.R
\name{expected_risk}
\alias{expected_risk}
\title{Calculate the expected risk for a portfolio}
\usage{
expected_risk(allocation)
}
\arguments{
\item{allocation}{The asset allocation of a portfolio}
}
\value{
a numeric
}
\description{
Calculate the expected risk for a portfolio
}
\examples{
x <- "Market Series ETF 070"
y <- get_portfolio_allocation(x)
exp_return <- expected_risk(y)

}
