% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_sharpe.R
\name{expected_sharpe}
\alias{expected_sharpe}
\title{Calculate the expected Sharpe Raio of a portfolio}
\usage{
expected_sharpe(allocation)
}
\arguments{
\item{allocation}{The asset allocation of a portfolio.}
}
\value{
a numeric vector.
}
\description{
Calculate the expected Sharpe Raio of a portfolio
}
\examples{
x <- get_portfolio_allocation("Multifactor Series ETF 080")
sharpe <- expected_sharpe(x)

}
