% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_yield.R
\name{expected_yield}
\alias{expected_yield}
\title{Calculate the expected average yield for a portfolio}
\usage{
expected_yield(allocation)
}
\arguments{
\item{allocation}{The asset class allocation of a portfolio.}
}
\value{
numeric
}
\description{
Calculate the expected average yield for a portfolio
}
\examples{
x <- "Market Series ETF 070"
y <- get_portfolio_allocation(x)
expected_yield(y)
}
