% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_long_short_products.R
\name{get_long_short_products}
\alias{get_long_short_products}
\title{Get Long/Short Products from S3}
\usage{
get_long_short_products()
}
\value{
A data frame containing product information (Product ID, Product Name, Ticker)
}
\description{
Retrieves the long/short products CSV file from AWS S3.
}
\examples{
\dontrun{
products <- get_long_short_products()
}
}
