% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_aggregates.R
\name{get_model_aggregates}
\alias{get_model_aggregates}
\title{Get Model Aggregates from S3}
\usage{
get_model_aggregates()
}
\value{
A data frame containing model aggregate information (modelAggId, modelName)
}
\description{
Retrieves the model aggregates CSV file from AWS S3.
}
\examples{
\dontrun{
model_aggs <- get_model_aggregates()
}
}
