% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_portfolio_allocation.R
\name{get_portfolio_allocation}
\alias{get_portfolio_allocation}
\title{Get the asset allocation of a model portfolio}
\usage{
get_portfolio_allocation(model_name)
}
\arguments{
\item{model_name}{A character vector with, at most, one element.}
}
\value{
A tibble.
}
\description{
Get the asset allocation of a model portfolio
}
\examples{
x <- "Multifactor Series ETF 060"
get_portfolio_allocation(x)

}
