% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_product_classification_framework.R
\name{get_product_classification_framework}
\alias{get_product_classification_framework}
\title{Get Product Classification Framework from S3}
\usage{
get_product_classification_framework()
}
\value{
A data frame containing product classification framework information
}
\description{
Retrieves the MA Product Classification Framework CSV file from AWS S3.
}
\examples{
\dontrun{
framework <- get_product_classification_framework()
}
}
