% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdot-package.R
\docType{package}
\name{kdot-package}
\alias{kdot}
\alias{kdot-package}
\title{kdot: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mercer-pmg/kdot}
  \item Report bugs at \url{https://github.com/mercer-pmg/kdot/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
