% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_returns.R
\name{monthly_returns}
\alias{monthly_returns}
\title{Calculate calendar month returns of a security}
\usage{
monthly_returns(x, start.date, end.date = Sys.Date())
}
\arguments{
\item{x}{a Bloomberg security id}

\item{start.date}{a date in YYYY-MM-DD format}

\item{end.date}{a date in YYYY-MM-DD format}
}
\value{
a tibble
}
\description{
Calculate calendar month returns of a security
}
