% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orion_product_import.R
\name{orion_product_import}
\alias{orion_product_import}
\title{Create Orion Product Local Update import}
\usage{
orion_product_import(sheet_with_assignments)
}
\arguments{
\item{sheet_with_assignments}{filename of Product Classification workbook}
}
\value{
a tibble
}
\description{
Create Orion Product Local Update import
}
