% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_assets_response.R
\name{parse_assets_response}
\alias{parse_assets_response}
\title{Parse Assets Response Data}
\usage{
parse_assets_response(assets_data)
}
\arguments{
\item{assets_data}{The assets data to parse (data.frame or list)}
}
\value{
A list of asset objects
}
\description{
Converts assets response data to a consistent list format.
}
\keyword{internal}
