% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_quilt.R
\name{performance_quilt}
\alias{performance_quilt}
\title{Create security performance quilt in XLSX workbook}
\usage{
performance_quilt(
  wb,
  sheetName,
  indexes,
  frequency,
  start.date,
  end.date = Sys.Date(),
  summary = TRUE,
  vol_frequency = "match",
  n = NULL,
  first_col = "latest",
  startRow = 2,
  startCol = 2,
  title = NULL,
  labels = NULL,
  palette = NULL
)
}
\arguments{
\item{wb}{XLSX workbook object}

\item{sheetName}{XSLX workbook sheet name.}

\item{indexes}{character vector of Bloomberg security IDs}

\item{frequency}{interval of performance calculation}

\item{start.date}{YYYY-MM-DD}

\item{end.date}{YYYY-MM-DD}

\item{summary}{If TRUE, the default displays summary of performance. If TRUE, no summary is displayed.}

\item{vol_frequency}{The return periods to use to calculated annualized volatility in the summary. By default matches the frequency of the quilt body.}

\item{n}{periods of performance to display}

\item{first_col}{"earliest" or "latest"}

\item{startRow}{row to place quilt}

\item{startCol}{column to place quilt}

\item{title}{title of performance quilt}

\item{labels}{labels to use inside quilt squares}

\item{palette}{fill colors to use in the quilt. Use Hex codes.}
}
\value{
a XLSX workbook object
}
\description{
Create security performance quilt in XLSX workbook
}
