% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_holdings.R
\name{portfolio_holdings}
\alias{portfolio_holdings}
\title{Get security-level portfolio holdings.}
\usage{
portfolio_holdings(strategy_name, date = Sys.Date())
}
\arguments{
\item{strategy_name}{character}

\item{date}{character}
}
\value{
a tibble
}
\description{
Get security-level portfolio holdings.
}
