% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_holdings_history.R
\name{portfolio_holdings_history}
\alias{portfolio_holdings_history}
\title{Get a portfolio's full holdings history}
\usage{
portfolio_holdings_history(portfolio_name)
}
\arguments{
\item{portfolio_name}{character}
}
\value{
tibble
}
\description{
Get a portfolio's full holdings history
}
\examples{
x <- "Multifactor Series ETF 060"
portfolio_holdings_history(x)
}
