% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_upload_factset.R
\name{portfolio_upload_factset}
\alias{portfolio_upload_factset}
\title{Create portfolio to be uploaded to FactSet}
\usage{
portfolio_upload_factset(portfolio_name)
}
\arguments{
\item{portfolio_name}{a character}
}
\value{
a tibble.
}
\description{
Create portfolio to be uploaded to FactSet
}
