% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_upload_ycharts.R
\name{portfolio_upload_ycharts}
\alias{portfolio_upload_ycharts}
\title{Create portfolio to be uploaded to Y Charts}
\usage{
portfolio_upload_ycharts(portfolio_name)
}
\arguments{
\item{portfolio_name}{character}
}
\value{
a tibble
}
\description{
Create portfolio to be uploaded to Y Charts
}
