% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarterly_returns.R
\name{quarterly_returns}
\alias{quarterly_returns}
\title{Calculate calendar quarter returns of a security}
\usage{
quarterly_returns(x, start.date, end.date = Sys.Date())
}
\arguments{
\item{x}{a Bloomberg security id}

\item{start.date}{a date in YYYY-MM-DD format}

\item{end.date}{a date in YYYY-MM-DD format}
}
\value{
a tibble
}
\description{
Calculate calendar quarter returns of a security
}
