% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_return.R
\name{strategy_return}
\alias{strategy_return}
\title{Calculate the return (net or gross) of a MA strategy}
\usage{
strategy_return(strategy_name, start, end, type = "net")
}
\arguments{
\item{strategy_name}{the name of an MA strategy}

\item{start}{yyyy-mm-dd}

\item{end}{yyyy-mm-dd}

\item{type}{either "net" or "gross"}
}
\value{
a numeric
}
\description{
Calculate the return (net or gross) of a MA strategy
}
