% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_return_mtd.R
\name{strategy_return_mtd}
\alias{strategy_return_mtd}
\title{MTD strategy return}
\usage{
strategy_return_mtd(strategy_name, type = "net")
}
\arguments{
\item{strategy_name}{the name of an MA strategy}

\item{type}{"gross" or "net" for fees}
}
\value{
a numeric
}
\description{
MTD strategy return
}
