% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_return_ytd.R
\name{strategy_return_ytd}
\alias{strategy_return_ytd}
\title{YTD strategy return}
\usage{
strategy_return_ytd(strategy_name, type = "net")
}
\arguments{
\item{strategy_name}{the name of an MA strategy}

\item{type}{"gross" or "net" for fees}
}
\value{
a numeric
}
\description{
YTD strategy return
}
