% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekly_returns.R
\name{weekly_returns}
\alias{weekly_returns}
\title{Calculate calendar week returns of a security}
\usage{
weekly_returns(x, start.date, end.date = Sys.Date())
}
\arguments{
\item{x}{A Bloomberg security id.}

\item{start.date}{A date.}

\item{end.date}{A date.}
}
\value{
A tibble.
}
\description{
Calculate calendar week returns of a security
}
