% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_orion_platform.R
\name{clean_orion_platform}
\alias{clean_orion_platform}
\title{Clean and Categorize Orion Platform Data}
\usage{
clean_orion_platform(data)
}
\arguments{
\item{data}{Orion Platform XXXX-XX.csv}
}
\value{
A tibble with added columns: 'category', 'model_series', 'asset_category', 'market_cap', and 'is_SMA'.
}
\description{
Clean and Categorize Orion Platform Data
}
