% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_orion_platform.R
\name{clean_orion_platform}
\alias{clean_orion_platform}
\title{Clean and Categorize Orion Platform Data}
\usage{
clean_orion_platform(data, drop_columns = NULL)
}
\arguments{
\item{data}{Orion Platform XXXX-XX.csv}

\item{drop_columns}{Character vector of column names to drop from the result (default: NULL)}
}
\value{
A tibble with added columns: 'category', 'model_group', 'asset_category', 'market_cap', and 'is_SMA'.
}
\description{
Clean and Categorize Orion Platform Data
}
