% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_classification_wb.R
\name{create_classification_wb}
\alias{create_classification_wb}
\title{Create a classification workbook to audit Orion products.}
\usage{
create_classification_wb(all_local)
}
\arguments{
\item{all_local}{an XLSX exported from Orion Local Products page.}
}
\value{
a list
}
\description{
Create a classification workbook to audit Orion products.
}
