% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_holdings.R
\name{portfolio_holdings}
\alias{portfolio_holdings}
\title{Get security-level portfolio holdings.}
\usage{
portfolio_holdings(portfolio_name, date = Sys.Date())
}
\arguments{
\item{portfolio_name}{character}

\item{date}{character}
}
\value{
a tibble
}
\description{
Get security-level portfolio holdings.
}
\examples{
x <- "Multifactor Series ETF 060"
portfolio_holdings(x)
portfolio_holdings(x, "2022-01-23")

}
